backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.40.0
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.4
